+++ ipm.py
@@ -449,7 +449,8 @@ class Thresholds:
             get thr all                     : List all existing thresholds in the subscription.
             get thr <product_code>          : List all thresholds of a particular product code (ex: lz, nt, etc)
             get thr <thr_name>              : Displays a single threshold in JSON format.
-            get thr -f <threshold_list>     : Displays multiple thresholds from a list in JSON format.
+            get thr -f  <threshold_list>    : Displays multiple thresholds from a list in JSON format.
+            get thr -rg <rg_id>             : Displays all the thresholds assigned to this Resource Group.
     """)
         sys.exit(1)
 
@@ -594,6 +595,51 @@ class Thresholds:
                         sys.exit(1)
                     sys.exit(0)
 
+                # Prints all existing thresholds assigned to a specific Resource Group
+                elif arguments[3] == '-rg':
+                    if (len(arguments) != 5):
+                        Thresholds.get_usage()
+                    
+                    rg_id = arguments[4]
+                    payload = '/1.0/thresholdmgmt/resource_assignments?_offset=1&_limit=99999'
+                    r = Thresholds.make_assignment_request(ipm_type,session_type,payload,subscription,region,username,password)
+
+                    if (r.status_code == 200):
+                        json_thr_assignments = json.loads(r.content)
+                    else:
+                        print ("Resource Group \'%s\' is invalid or empty. Exiting!" % rg_id)
+                        sys.exit(1)
+                    
+                    n = 0
+                    matches = 0
+                    count_of_rgs = len(json_thr_assignments['_items'])
+                    
+                    if (count_of_rgs > 0):
+                        
+                        for _ in json_thr_assignments['_items']:
+                            
+                            if (rg_id in (json_thr_assignments['_items'][n]['resource']['_id'])):
+                                matches += 1
+                                if (matches == 1):
+                                    print ("'threshold_name','product_code','threshold_type','description'")
+                                
+                                payload = json_thr_assignments['_items'][n]['threshold']['_href']
+                                r = Thresholds.make_assignment_request(ipm_type,session_type,payload,subscription,region,username,password)
+
+                                if (r.status_code == 200):
+                                    thresholds_dic = json.loads(r.content)
+                                else:
+                                    print ("ERROR - Failed to complete request on given item. Exiting!")
+                                    sys.exit(1)
+                                
+                                print ("'" + thresholds_dic['label'] + "','" + thresholds_dic['_appliesToAgentType'] + "','" + thresholds_dic['description'] + "'")
+
+                            n += 1
+                        if (matches == 0):
+                            print ("ERROR - Resource Group '%s' is invalid or doesn't have any resources assigned." %(rg_id))
+                            sys.exit(1)
+                    sys.exit(0)
+
                 # prints a single threshold that was informed
                 threshold_name = arguments[3]
                 payload = '/1.0/thresholdmgmt/threshold_types/itm_private_situation/thresholds/?_filter=label%3D' + threshold_name
@@ -670,6 +716,25 @@ class Thresholds:
         else:
             print ("ERROR - Could not determine IPM subscription type. Exiting!")
             sys.exit(1)
+
+    @staticmethod
+    def make_assignment_request(ipm_type,session_type,payload,subscription,region,username,password):
+        """ GETs list of Thresholds assigned to a specific threshold."""
+
+
+        if (ipm_type == "cloud"):
+            url = 'https://' + subscription + '.customers.' + region + '.apm.ibmserviceengage.com' + payload
+            r = requests.get(url, auth=(username, password), timeout=60)
+            return r
+        elif (ipm_type == "private"):
+            url = 'https://' + subscription + payload
+            requests.packages.urllib3.disable_warnings(InsecureRequestWarning)
+            r = requests.get(url, auth=(username, password), verify=False, timeout=60)
+            return r
+        else:
+            print ("ERROR - Could not determine IPM subscription type. Exiting!")
+            sys.exit(1)
+
 
     @staticmethod
     def if_not_empty(value):
@@ -1052,7 +1117,8 @@ ipm get <object> / <object_id>
     get agt                             : List all existing agents on the subscription.
     get thr                             : List of all available thresholds.
     get thr <thr_name>                  : Displays a single threshold in JSON format.
-    get thr -f <threshold_list>         : Displays multiple thresholds from a list in JSON format.
+    get thr -f  <threshold_list>        : Displays multiple thresholds from a list in JSON format.
+    get thr -rg <rg_id>                 : Displays all the thresholds assigned to this Resource Group.
     get rg                              : List of all available Resource Groups.
     get rg <rg_id>                      : List of all Managed Systems assigned to this Resource Group.
 
@@ -1062,7 +1128,7 @@ ipm add <object> <object_id>
     add thr <threshold_json_file>       : Creates a threshold from an IPM8 JSON export file
 
 ipm del <object> <object_id>
-    del thr <threshold_name             : Deletes a threshold by name
+    del thr <threshold_name>            : Deletes a threshold by name
     del rg  <resourcegroup_id>          : Deletes a Resource Group by Id
     del agt <agt_name> <rg_id>          : Removes an agent from a Resource Group